Program

+++ MonitoringAndEvaluationPlatform/Models/SubOutput.cs
@@ -8,10 +8,10 @@ namespace MonitoringAndEvaluationPlatform.Models
         public int Code { get; set; }
         public string Name { get; set; }
         public double Trend { get; set; } = 0;
-        public int IndicatorsPerformance { get; set; } = 0;
-        public int DisbursementPerformance { get; set; } = 0;
-        public int FieldMonitoring { get; set; } = 0;
-        public int ImpactAssessment { get; set; } = 0;
+        public double IndicatorsPerformance { get; set; } = 0;
+        public double DisbursementPerformance { get; set; } = 0;
+        public double FieldMonitoring { get; set; } = 0;
+        public double ImpactAssessment { get; set; } = 0;
 
         [Range(0, 1, ErrorMessage = "The Weight must be between 0 and 1.")]
         public double Weight { get; set; } = 1;
